chore: Move  family model to families app
