+++ accounts/models/__init__.py
@@ -1,3 +1,2 @@
-from accounts.models.family import Family
 from accounts.models.user import User
