+++ accounts/models/user.py
@@ -1,9 +1,10 @@
 from django.db import models
 from django.contrib.auth.models import AbstractUser
+from families.models import Family
 
 
 class User(AbstractUser):
-    family_id = models.ForeignKey('Family', on_delete=models.CASCADE, null=True, blank=True)
+    family_id = models.ForeignKey(Family, on_delete=models.CASCADE, null=True, blank=True)
     member_id = models.CharField(max_length=31, null=True, default='')
     password = models.CharField(max_length=31, null=True, default='')
     name = models.CharField(max_length=31, null=True, default='')
