+++ archives/models/album.py
@@ -1,7 +1,7 @@
 from django.db import models
 
 from archives.models.base_model import BaseModel
-from accounts.models.family import Family
+from families.models import Family
 
 class Album(BaseModel):
     family_id = models.ForeignKey(Family, on_delete=models.CASCADE, null=True, blank=True)
