+++ archives/models/photo.py
@@ -2,7 +2,8 @@ from django.db import models
 
 from archives.models.base_model import BaseModel
 from archives.models.album import Album
-from accounts.models.family import Family
+from families.models import Family
+
 
 
 class Photo(BaseModel):
