+++ archives/views.py
@@ -2,7 +2,7 @@ import json
 
 from django.http import JsonResponse, HttpResponse
 from django.shortcuts import get_object_or_404
-from accounts.models import Family
+from families.models import Family
 from .models import Album, Like, Photo
 
 from django.views.decorators.http import require_http_methods
