+++ families/admin.py
@@ -1,7 +1,6 @@
 from django.contrib import admin
 
-from .models import User
-
 # Register your models here.
-admin.site.register(User)
+from .models import Family
 
+admin.site.register(Family)
